+++ src/service/api-frontend.ts
@@ -5,12 +5,26 @@ import shortid from 'shortid';
 const mockToken = 'testabc.xyz.ahk';
 
 class ApiFrontend extends IAPI {
+  isLoggedIn(): boolean {
+    const token = localStorage.getItem('token');
+
+    return !!token;
+  }
+
   async signIn(username: string, password: string): Promise<string> {
     if (username === 'firstUser' && password === 'example') {
+      localStorage.setItem('token', mockToken);
+
       return Promise.resolve(mockToken);
     }
 
     return Promise.reject('Incorrect username/password');
+  }
+
+  async signOut() {
+    localStorage.removeItem('token');
+
+    return Promise.resolve();
   }
 
   async createTodo(content: string): Promise<Todo> {
@@ -24,7 +38,18 @@ class ApiFrontend extends IAPI {
   }
 
   async getTodos(): Promise<Todo[]> {
-    return [];
+    const json = localStorage.getItem('todoList');
+    let result = [];
+
+    if (json) {
+      try {
+        result = JSON.parse(json).map((item: any) => item as Todo);
+      } catch ({message}) {
+        console.error(message);
+      }
+    }
+
+    return Promise.resolve(result);
   }
 }
