+++ src/service/index.ts
@@ -3,9 +3,9 @@ import {IAPI} from './types';
 let Service: IAPI;
 
 if (process.env.REACT_APP_WHOAMI === 'frontend') {
-  Service = require('./api-frontend.service').default as IAPI;
+  Service = require('./api-frontend').default as IAPI;
 } else {
-  Service = require('./api-fullstack.service').default as IAPI;
+  Service = require('./api-fullstack').default as IAPI;
 }
 
 export default Service;
